forwarder/pyproject.toml
[tool.poetry]
name = "forwarder"
version = "0.1.0"
description = ""
authors = ["Bob Steers <steers.bob@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
paho-mqtt = "^1.5.1"
click = "^7.1.2"

[tool.poetry.dev-dependencies]
pytest-flake8 = "^1.0.6"
pytest-cov = "^2.10.1"
pytest-mock = "^3.3.1"
flake8 = "^3.8.4"
autopep8 = "^1.5.4"
pytest = "^6.1.2"
flake8-quotes = "^3.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
